Allow reading and writing out of the bounds of the canvas
Reading a dot off of the canvas will always return `False` and writing
to a dot off of the canvas will always act as a no-op. This change will
allow for the implementation of drawing algorithms using simple code
that does not have to worry about bounds checking every canvas dot
access.