Replace tuple coordinates with individual int parameters
Originally a two-int-tuple was going to represent the fundamental
position unit within the library. However, with the addition of the
Point class it is somewhat confusing to have a point representations and
a position-tuple representation that are similar, but not quite equal.

This changeset replaces Canvas method tuple parameters with individual
arguments for x and y (or width and height in the case of __init__).